+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v2.1.0
 
+- New feature: Export
+  - You can now export your soul break inventory and legend materia inventory in CSV format.
+  - Exporting all of RK Squared's state as JSON is also supported.
 - Options
   - Disable the static battle background options - It no longer works in current versions of FFRK. In practice, it was of little benefit; enemy special effects and soul break animations hurt performance worse.